# docs/swagger.yaml
definitions:
...
  dtos.MessagesResponses:
    properties:
      message:
        type: string
    type: object
...
  dtos.Product:
    properties:
      active:
        type: boolean
      buy_price:
        type: integer
      category_id:
        type: integer
      defective:
        type: integer
      min_retail_price:
        type: integer
      picture:
        type: string
      position:
        type: string
      product_code:
        type: string
      product_name:
        type: string
      quantity:
        type: integer
      size_type_id:
        type: integer
    type: object
  dtos.SizeType:
    properties:
      name:
        type: string
    type: object
...
info:
  contact:
    email: cs@payuoge.com
    name: API Support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /groceries/products:
    get:
      consumes:
      - application/json
      description: do get all product
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: GetAll product groceries process
      tags:
      - groceries
    post:
      consumes:
      - application/json
      description: do create a product
      parameters:
      - description: create a product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dtos.Product'
      produces:
      - application/json
      responses:
        "200":
          description: the message successfully create
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad Request
          schema:
            type: string
      security:
      - Bearer: []
      summary: CreateProduct access process
      tags:
      - groceries
...
  /product/{id}:
    get:
      consumes:
      - application/json
      description: do get id product
      parameters:
      - description: get id product
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: get id category
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: cookies not found
          schema:
            type: string
      security:
      - Bearer: []
      summary: GetID product process
      tags:
      - products
  /products:
    get:
      consumes:
      - application/json
      description: do get all product
      produces:
      - application/json
      responses:
        "200":
          description: get all category
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: cookies not found
          schema:
            type: string
      security:
      - Bearer: []
      summary: GetAll product process
      tags:
      - products
...
  /products/size:
    get:
      consumes:
      - application/json
      description: do get all size
      produces:
      - application/json
      responses:
        "200":
          description: get all size
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: cookie not found
          schema:
            type: string
      security:
      - Bearer: []
      summary: GetAll Size process
      tags:
      - products
    post:
      consumes:
      - application/json
      description: do create size type
      parameters:
      - description: create sizetype
        in: body
        name: size
        required: true
        schema:
          $ref: '#/definitions/dtos.SizeType'
      produces:
      - application/json
      responses:
        "200":
          description: create size type successfully
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: Error Bad request
          schema:
            type: string
      security:
      - Bearer: []
      summary: create sizeType process
      tags:
      - products
  /products/size/{id}:
    delete:
      consumes:
      - application/json
      description: do delete a size
      parameters:
      - description: delete size with id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: get all size
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: cookie not found
          schema:
            type: string
      security:
      - Bearer: []
      summary: DeleteID Size process
      tags:
      - products
    get:
      consumes:
      - application/json
      description: do get all size
      parameters:
      - description: get id size
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: get all size
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: cookie not found
          schema:
            type: string
      security:
      - Bearer: []
      summary: GetID Size process
      tags:
      - products
    put:
      consumes:
      - application/json
      description: do get all size
      parameters:
      - description: access id size type
        in: path
        name: id
        required: true
        type: integer
      - description: update size type
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/dtos.SizeType'
      produces:
      - application/json
      responses:
        "200":
          description: get all size
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: cookie not found
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update Size process
      tags:
      - products
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"